+++ src/test/java/testsupport/EquatableM.java
@@ -16,6 +16,14 @@ public final class EquatableM<M extends Monad<?, M>, A> implements Monad<A, Equa
         this.equatable = equatable;
     }
 
+    public <B> EquatableM<M, B> with(Fn1<? super Monad<A, M>, ? extends Monad<B, M>> fn) {
+        return new EquatableM<>(fn.apply(ma), equatable);
+    }
+
+    public <B> EquatableM<M, B> swap(Monad<B, M> mb) {
+        return new EquatableM<>(mb, equatable);
+    }
+
     @Override
     public <B> EquatableM<M, B> flatMap(Fn1<? super A, ? extends Monad<B, EquatableM<M, ?>>> f) {
         return new EquatableM<>(ma.flatMap(f.fmap(x -> x.<EquatableM<M, B>>coerce().ma)), equatable);